backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Background Jobs
celery==5.3.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4

# SSO
python3-saml==1.16.0
authlib==1.3.0

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Encryption
cryptography==42.0.0

# HTTP Client
httpx==0.26.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0.post1

# Logging
loguru==0.7.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# WebSocket
websockets==12.0
